feat(SDL): migrate SDL3 code from develop

+++ src/custom_sprites.c
@@ -41,7 +41,7 @@
 // #define OUTER
 // #define INNER
 #if defined(OUTER) || defined(INNER)
-#include <SDL2/SDL.h>
+#include <SDL3/SDL.h>
 #endif
 
 // Each part of RGB tuple of palette file is 1-63 actually
@@ -1203,7 +1203,7 @@ collect_sprites(const char *path, unzFile zip, const char *blender_scene, struct
                     store_ksp_fc = context->ksp_first->FramesCount;
 #ifdef INNER
                 fprintf(stderr, "F:%s/%s\n", path, name);
-                fprintf(stderr, "A:%d\n", SDL_GetTicks());
+                fprintf(stderr, "A:%u\n", (unsigned)SDL_GetTicks());
 #endif
                 if (!read_png_data(zip, path, context, name, is_fp, node, itm))
                 {
@@ -1219,7 +1219,7 @@ collect_sprites(const char *path, unzFile zip, const char *blender_scene, struct
                     return 1;
                 }
 #ifdef INNER
-                fprintf(stderr, "B:%d\n", SDL_GetTicks());
+                fprintf(stderr, "B:%u\n", (unsigned)SDL_GetTicks());
 #endif
                 if (UNZ_OK != unzCloseCurrentFile(zip))
                 {
